fix: resolve ESLint errors in frontend for CI pass
- Split AuthContext into separate files to fix react-refresh/only-export-components
- Replace `any` with `unknown` in DataTable generic constraint
- Remove unused catch variables in ChatPage, IntegrationsPage, LoginPage
