+++ platform/src/hooks/auth-context.ts
@@ -1,0 +1,10 @@
+import { createContext } from 'react'
+
+export interface AuthContextType {
+  token: string | null
+  isAuthenticated: boolean
+  login: (token: string) => void
+  logout: () => void
+}
+
+export const AuthContext = createContext<AuthContextType | undefined>(undefined)
